HW_Report/week13/requirements.txt
anyio==4.9.0
asgiref==3.8.1
channels==4.2.2
click==8.1.8
Django==5.2
djangorestframework==3.16.0
h11==0.16.0
httptools==0.6.4
idna==3.10
python-dotenv==1.1.0
PyYAML==6.0.2
requests==2.31.0
sniffio==1.3.1
sqlparse==0.5.3
typing_extensions==4.13.2
uvicorn==0.34.2
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
gunicorn
psycopg2-binary
channels
whitenoise
django-allauth==0.61.0